refactor(tests): move matchesAnyFilterPattern unit tests into ical.test.ts
Consolidates all ical-related tests into one file, following existing
repo convention. Removes the separate calendar-filter.test.ts file.
